ark_signv2/requirements.txt
multidict==6.0.3
yarl==1.8.2
async_timeout==4.0.2
charset_normalizer==3.0.1
aiosignal==1.3.1
uvicorn==0.19.0
aiofiles==22.1.0
aiohttp==3.8.3
fastapi==0.87.0
slowapi==0.1.7
aioredis==2.0.1